[robot-descriptions] Renaming of imports
