+++ src/pycram/robot_descriptions/stretch_description.py
@@ -1,6 +1,6 @@
 import rospkg
 
-from ..new_robot_description import RobotDescription, KinematicChainDescription, EndEffectorDescription, \
+from ..robot_description import RobotDescription, KinematicChainDescription, EndEffectorDescription, \
     CameraDescription, RobotDescriptionManager
 
 rospack = rospkg.RosPack()
